[CHORE] : move mock data to HomeViewModel.kt
